add error messages for not found exe for building

+++ get-dependencies.ps1
@@ -6,6 +6,14 @@ param(
     $python_exe_path = "$($HOME)\AppData\Local\Programs\Python\Python39\python.exe"
 )
 
+if (!(Test-Path $python_exe_path)) {
+    throw "python exe path not found please update to your path to python";
+}
+
+if (!(Test-Path $vcpkg_exe_path)) {
+    throw "vcpkg exe path not found please update to your path to vcpkg and/or install vcpkg first";
+}
+
 & $vcpkg_exe_path install --triplet x64-windows boost-algorithm boost-date-time boost-endian boost-functional boost-iostreams boost-rational boost-signals2 boost-stacktrace doctest minizip nlohmann-json pugixml;
 
 & $python_exe_path -m pip install setuptools wheel py7zr==0.20.*;
